Add navigation and book functions modules
