+++ index.js
@@ -1,63 +1,6 @@
 import { DateTime } from './modules/luxon.js';
-
-class Book {
-    constructor(title, author) {
-      this.title = title;
-      this.author = author;
-    }
-  }
- 
-  if (localStorage.getItem('Book List') === null) {
-    localStorage.setItem('Book List', JSON.stringify([]));
-  }
-  
-  const bookList = JSON.parse(localStorage.getItem('Book List'));
-  
-  function updateLocalStorage() {
-    localStorage.setItem('Book List', JSON.stringify(bookList));
-  }
-  
-  const displayBooks = () => {
-    const booksContainer = document.querySelector('.books-container');
-    booksContainer.innerHTML = '';
-    for (let i = 0; i < bookList.length; i += 1) {
-      if ((i % 2) === 0) {
-        booksContainer.innerHTML += `<div class="book-data grey"><ul>
-        <li>"${bookList[i].title}" by ${bookList[i].author}</li>
-        </ul>
-        <button class="remove-button" onclick="removeBook(${i})">Remove</button></div>`;
-      } else {
-        booksContainer.innerHTML += `<div class="book-data white"><ul>
-        <li>"${bookList[i].title}" by ${bookList[i].author}</li>
-        </ul>
-        <button class="remove-button" onclick="removeBook(${i})">Remove</button></div>`;
-      }
-    }
-  };
-  
-  const clearInput = () => {
-    document.querySelector('.book-title').value = '';
-    document.querySelector('.book-author').value = '';
-  };
-  const addBook = (bookTitle, bookAuthor) => {
-    if (bookTitle === '' || bookAuthor === '') {
-      alert('Fields cannot be empty');
-    } else {
-      const books = new Book(bookTitle, bookAuthor);
-      bookList.push(books);
-      updateLocalStorage();
-      displayBooks();
-      clearInput();
-    }
-  };
-  
-  // eslint-disable-next-line
-  const removeBook = (i) => {
-    bookList.splice(i, 1);
-    updateLocalStorage();
-    displayBooks();
-  };
-  
+import { removeBook, addBook, clearInput,  displayBooks} from './modules/bookFunctions.js';
+import { listMenu, contactMenu, addNewMenu  } from './modules/navigation.js'
   const form = document.querySelector('form');
   form.addEventListener('submit', (e) => {
     const bookTitle = document.querySelector('.book-title');
@@ -68,34 +11,22 @@ class Book {
   
   window.onload = displayBooks();
   
-  const bookContainerTitle = document.querySelector('.title');
-  const bookContainer = document.querySelector('.books-container');
-  const addBookContainer = document.querySelector('.add-book-container');
-  const contactContainer = document.querySelector('.contact-container');
   const list = document.querySelector('#list');
   const addNew = document.querySelector('#addNew');
   const contact = document.querySelector('#contact');
   
   list.addEventListener('click', () => {
-    bookContainerTitle.style.display = 'block';
-    bookContainer.style.display = 'block';
-    contactContainer.style.display = 'none';
-    addBookContainer.style.display = 'none';
+    listMenu();
   });
   
   addNew.addEventListener('click', () => {
-    bookContainerTitle.style.display = 'none';
-    bookContainer.style.display = 'none';
-    contactContainer.style.display = 'none';
-    addBookContainer.style.display = 'block';
+    addNewMenu();
   });
   
   contact.addEventListener('click', () => {
-    bookContainerTitle.style.display = 'none';
-    bookContainer.style.display = 'none';
-    contactContainer.style.display = 'flex';
-    addBookContainer.style.display = 'none';
+    contactMenu();
   });
 
   const Date = document.querySelector('.date-time');
+
   setInterval(function () {Date.innerHTML = `${DateTime.local().toLocaleString(DateTime.DATETIME_MED_WITH_SECONDS)}`;}, 1000);
